Move the isNewlyPublished state from UrlState to AppState
The is newly published was in the URL state for when we popped open
a new tab to display the profile. This is no longer needed in the
URL, but is still in the store to show the permalink.
