+++ src/reducers/url-state.js
@@ -370,21 +370,6 @@ const profileName: Reducer<string> = (state = '', action) => {
 };
 
 /**
- * This reducer holds the state for whether or not a profile was newly uploaded
- * or not. This way we can show a friendly message to the user.
- */
-const isNewlyPublished: Reducer<boolean> = (state = false, action) => {
-  switch (action.type) {
-    case 'PROFILE_PUBLISHED':
-      return true;
-    case 'DISMISS_NEWLY_PUBLISHED':
-      return false;
-    default:
-      return state;
-  }
-};
-
-/**
  * These values are specific to an individual profile.
  */
 const profileSpecific = combineReducers({
@@ -449,7 +434,6 @@ const urlStateReducer: Reducer<UrlState> = wrapReducerInResetter(
     pathInZipFile,
     profileSpecific,
     profileName,
-    isNewlyPublished,
   })
 );
