+++ src/test/components/MenuButtonsPermalinks.test.js
@@ -10,7 +10,6 @@ import { Provider } from 'react-redux';
 import { storeWithProfile } from '../fixtures/stores';
 import { stateFromLocation } from '../../app-logic/url-handling';
 import { ensureExists } from '../../utils/flow';
-import { getIsNewlyPublished } from '../../selectors/url-state';
 
 describe('<Permalink>', function() {
   function setup(search = '', injectedUrlShortener) {
@@ -73,21 +72,5 @@ describe('<Permalink>', function() {
       'Unable to find the permalink input text field'
     );
     expect(input.getAttribute('value')).toBe(shortUrl);
-  });
-
-  it('shows the permalink when visiting a published profile', async function() {
-    const { queryInput, shortUrl, shortUrlPromise } = setup('?published');
-    await shortUrlPromise;
-    const input = ensureExists(
-      queryInput(),
-      'Unable to find the permalink input text field'
-    );
-    expect(input.getAttribute('value')).toBe(shortUrl);
-  });
-
-  it('resets the isNewlyPublishedState in the URL when mounted', async function() {
-    const { originalState, getState } = setup('?published');
-    expect(getIsNewlyPublished(originalState)).toBe(true);
-    expect(getIsNewlyPublished(getState())).toBe(false);
   });
 });
